requirements.txt
matplotlib
numpy 
pandas 
scipy 
scikit-learn 
LightGBM 
CatBoost
TensorFlow 
PyTorch 
seaborn 
statsmodels 
